Correção dos metodos e alteração do data transfer para a nova maneira de realizar as transferencias e consulta dos saldos

+++ src/main/java/com/desafio/apitransferencia/dto/TransacaoDTO.java
@@ -2,5 +2,5 @@ package com.desafio.apitransferencia.dto;
 
 import java.math.BigDecimal;
 
-public record TransacaoDTO(BigDecimal valorTransacao, int numeroContaPagador, int numeroContaRecebedor) {
+public record TransacaoDTO(BigDecimal valorTransacao, long numeroContaPagador, long numeroContaRecebedor) {
 }
